wip(api-headless-cms): date storage transform
