+++ packages/api-headless-cms/src/features/storage/abstractions/StorageTransform.ts
@@ -27,5 +27,7 @@ export const StorageTransform = createAbstraction<IStorageTransform<any, any, an
 export namespace StorageTransform {
     export type Interface<T = any, R = any, F extends CmsModelField = CmsModelField> = IStorageTransform<T, R, F>;
     export type ToStorageParams<T = any, R = any, F extends CmsModelField = CmsModelField> = IStorageTransformToStorageParams<T, R, F>;
+    export type ToStorageResponse<T = any> = Promise<T>;
     export type FromStorageParams<T = any, R = any, F extends CmsModelField = CmsModelField> = IStorageTransformFromStorageParams<T, R, F>;
+    export type FromStorageResponse<T = any> = Promise<T>;
 }
