+++ packages/api-headless-cms/src/features/storage/feature.ts
@@ -1,9 +1,12 @@
 import { createFeature } from "@webiny/feature/api";
 import { StorageTransformRegistry } from "./StorageTransformRegistry.js";
+import { DateStorageTransform } from "./fields/DateStorageTransform.js";
 
 export const StorageFeature = createFeature({
     name: "Cms/StorageFeature",
     register: container => {
+        
+        container.register(DateStorageTransform);
         container.register(StorageTransformRegistry);
     }
 });
