+++ packages/api-headless-cms/src/plugins/index.ts
@@ -1,6 +1,5 @@
 export * from "./CmsGroupPlugin.js";
 export * from "./CmsModelPlugin.js";
-export * from "./StorageTransformPlugin.js";
 export * from "./CmsParametersPlugin.js";
 export * from "./CmsModelFieldConverterPlugin.js";
 export * from "./CmsGraphQLSchemaPlugin/index.js";
